requirements/prod.txt
Django==5.2.10
django-ckeditor-5==0.2.20
django-cleanup==9.0.0
django-environ==0.12.0
pillow==12.1.1
sorl-thumbnail==13.0.0